Fix image_operations.mix unit conversion bug for values exceeding 255. Fix povray bug that didn't write lights more than 1.

+++ makehuman/shared/image_operations.py
@@ -96,7 +96,9 @@ def blurred(img, level=10, kernelSize=10, progressCallback=None):
 def mix(img1, img2, weight1, weight2 = None):
     if weight2 is None:
         weight2 =  1 - weight1
-    return Image(data = numpy.around(weight1*img1.data + weight2*img2.data).astype(int))
+    return Image(data = numpy.around(
+        weight1*img1.data.astype(float)
+        + weight2*img2.data.astype(float)).astype(int))
 
 def clipped (img):
     return Image(data = numpy.clip(img.data,0,255))
